requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
requests>=2.31.0
apscheduler>=3.10.4
streamlit>=1.33.0
pandas>=2.2.0
psycopg2-binary>=2.9.9
pytest>=8.0.0